extract components from LoginFormModal to re-use in WAYF
